job-search-be/requirements.txt
# Base FastAPI requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# For LLM interaction
google-generativeai>=0.5.4
python-dotenv>=1.0.1

# For file conversion (DOCX to PDF)
docx2pdf>=0.1.8